backend/requirements.txt
django==3.2.25
djongo==1.3.6
pymongo==3.12.3
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
requests==2.32.3
pillow==10.4.0